backend/requirements.txt
Flask==2.3.2
pytube==15.0.0
Flask-Cors==3.0.10
python-dotenv==1.0.0